refactor: PlatformContributionExtractor 위치 변경 (#329)
