+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/user/infrastructure/calculator/GithubContributionCalculator.java
@@ -2,7 +2,7 @@ package com.woowacourse.pickgit.user.infrastructure.calculator;
 
 import com.woowacourse.pickgit.user.domain.Contribution;
 import com.woowacourse.pickgit.user.domain.PlatformContributionCalculator;
-import com.woowacourse.pickgit.user.domain.PlatformContributionExtractor;
+import com.woowacourse.pickgit.user.infrastructure.extractor.PlatformContributionExtractor;
 import com.woowacourse.pickgit.user.infrastructure.dto.CountDto;
 import com.woowacourse.pickgit.user.infrastructure.dto.ItemDto;
 import com.woowacourse.pickgit.user.infrastructure.dto.StarsDto;
